PobimSplatting/Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
werkzeug==3.0.1
numpy==1.26.4
pillow==10.1.0
opencv-python==4.8.1.78
trimesh==4.0.5
plyfile==1.0.2
gunicorn==21.2.0
eventlet==0.33.3
# Mesh conversion dependencies
pymeshlab==2023.12
scipy==1.11.4
pygltflib==1.16.1